fix: moved BackToTop logic into client component to avoid 'use client' conflict in layout
